refactor(core): add function maybeRepoNameFromSMFile
